move TrackingPointToCfgNode back to semanticcpg so it can be reused
